requirements.txt
appnope                0.1.3
argon2-cffi            21.3.0
argon2-cffi-bindings   21.2.0
asciitree              0.3.3
asttokens              2.0.8
attrs                  22.1.0
Babel                  2.10.3
backcall               0.2.0
beautifulsoup4         4.11.1
bleach                 5.0.1
certifi                2022.6.15
cffi                   1.15.1
charset-normalizer     2.1.1
click                  8.1.3
colorama               0.4.5
debugpy                1.6.3
decorator              5.1.1
defusedxml             0.7.1
distlib                0.3.4
docker-pycreds         0.4.0
docopt                 0.6.2
entrypoints            0.4
environs               9.5.0
et-xmlfile             1.1.0
executing              0.10.0
fastjsonschema         2.16.1
filelock               3.7.1
fire                   0.4.0
Flask                  2.2.2
Flask-Cors             3.0.10
gitdb                  4.0.9
GitPython              3.1.27
huggingface-hub        0.9.1
idna                   3.3
importlib-metadata     4.12.0
importlib-resources    5.9.0
install                1.3.5
interchange            2021.0.4
ipykernel              6.15.1
ipython                8.4.0
ipython-genutils       0.2.0
ipywidgets             8.0.1
itsdangerous           2.1.2
jarowinkler            1.2.1
jdcal                  1.4.1
jedi                   0.18.1
Jinja2                 3.1.2
joblib                 1.1.0
jsonlines              3.1.0
jsonschema             4.14.0
jupyter                1.0.0
jupyter-client         7.3.4
jupyter-console        6.4.4
jupyter-core           4.11.1
jupyterlab-pygments    0.2.2
jupyterlab-widgets     3.0.2
liac-arff              2.5.0
loguru                 0.6.0
lxml                   4.9.1
MarkupSafe             2.1.1
marshmallow            3.18.0
matplotlib-inline      0.1.6
minio                  7.1.12
mistune                2.0.4
mo-dots                4.22.21108
mo-future              3.147.20327
mo-imports             3.149.20327
mo-kwargs              4.22.21108
mo-logs                4.23.21108
monotonic              1.6
moz-sql-parser         4.40.21126
mysql-connector-python 8.0.30
nbclient               0.6.7
nbconvert              7.0.0
nbformat               5.4.0
neo4j                  5.0.0
neotime                1.7.4
nest-asyncio           1.5.5
nltk                   3.7
notebook               6.4.12
numpy                  1.23.2
openml                 0.12.2
openpyxl               2.4.11
packaging              21.3
pandas                 1.4.3
pandocfilters          1.5.0
pansi                  2020.7.3
parso                  0.8.3
pathtools              0.1.2
pexpect                4.8.0
pickleshare            0.7.5
pip                    22.2.2
pkgutil_resolve_name   1.3.10
platformdirs           2.5.2
prometheus-client      0.14.1
promise                2.3
prompt-toolkit         3.0.30
protobuf               3.20.1
psutil                 5.9.1
psycopg2-binary        2.9.3
ptyprocess             0.7.0
pure-eval              0.2.2
py2neo                 2021.2.3
pyarrow                9.0.0
pycparser              2.21
Pygments               2.13.0
pyparsing              3.0.9
pyrsistent             0.18.1
python-dateutil        2.8.2
python-dotenv          0.21.0
pytz                   2022.2.1
PyYAML                 6.0
pyzmq                  23.2.1
qtconsole              5.3.1
QtPy                   2.2.0
rapidfuzz              2.6.0
records                0.5.3
regex                  2022.8.17
requests               2.28.1
scikit-learn           1.1.2
scipy                  1.9.1
Send2Trash             1.8.0
sentry-sdk             1.9.5
setproctitle           1.3.2
setuptools             49.2.1
shortuuid              1.0.9
six                    1.15.0
smmap                  5.0.0
soupsieve              2.3.2.post1
SQLAlchemy             1.4.40
sqlparse               0.4.2
stack-data             0.4.0
tablib                 3.2.1
termcolor              2.0.1
terminado              0.15.0
threadpoolctl          3.1.0
tinycss2               1.1.1
tokenizers             0.12.1
torch                  1.12.1
tornado                6.2
tqdm                   4.64.0
traitlets              5.3.0
transformers           4.21.2
transtab               0.0.2
typing_extensions      4.3.0
urllib3                1.26.12
virtualenv             20.14.1
wandb                  0.13.2
wcwidth                0.2.5
webencodings           0.5.1
Werkzeug               2.2.2
wheel                  0.36.2
widgetsnbextension     4.0.2
xgboost                1.6.2
xmltodict              0.13.0
zipp                   3.8.1